Minor enhancements, B12 upload
B12 of is on the way of getting released on BukkitDev.
For a full changelog, visit CraftZ's Bukkit page or check the last 7
commits (and this one, of course).

In this commit, I corrected two formal things, but I also added the
option to turn death zombie spawning on and off.
In addition, I switched my Bukkit library over to a 1.7.5 Dev build, and
it seems that some more stuff will break very soon.

Anyway, enjoy playing CraftZ as much as I enjoy coding it!