Add purge command, small fixes
Since Bukkit is about to die apparently, I see no reason in holding this
commit back until something more is added (since it is possible that
nothing will be added for a long time). I just want to have this repo
up-to-date.
Changes:
- Small changes with the stack parser, mainly safety checks being added.
- A "/craftz purge" command was added to kill all zombies in the world.

I hope to continue this project at some point.