Tab completion, lobby & world border improvements
* All commands can be completed in chat by pressing TAB. This even works
for loot list names in the '/craftz sign' command.
* The help command output was improved. It is formatted better, and when
executed in the console, player-only commands are gray instead of
yellow.
* The world border can be configured with '/craftz setborder'. It can
have a square shape now as well, if you want (still defaults to round).
It has its own center, instead of using the lobby.
* Getting out of the world is a bit worse now regarding the potion
effects.
* '/craftz setlobby' now also configures the radius. The direction the
player is looking in is saved, too.