Add kits
This commit finally adds kits, I know you've been waiting for a long
time.
Two default kits are shipped with CraftZ - they can of course by removed
or modified as you like. The standard one contains no items, the other
one contains basic stuff.
When a player joins he receives the kit that is defined as the default
one, he can then select a different one with '/craftz kit <name>'.
The items are soulbound so it is impossible to generate lots and lots of
items by constantly spawning; when trying to drop them or when dying,
they immediately despawn.