Code style (lootchest and spawnpoint classes)
This commit doesn't add very much regarding features, but it should
improve the look of CraftZ's code. There is a new class,
WorldDataObject, that, together with its subclasses, forms another step
towards a more modular code that can be maintained easier.
It is used by the chest refiller, zombie spawner, and player manager's
spawn functions.

In addition, there was one bug fix: When the lobby was in a different
world and players tried to spawn, they had to enter the command two
times before actually spawning.