Use zombie metadata for storing player inventory
Instead of using the world data file to store the UUID of a zombie
together with the inventory contents and then later retrieving it when
the zombie dies, CraftZ now makes use of the metadata methods and
classes Bukkit provides.
This means the zombie entity itself is the one storing the stuff the
player had, which results in a lot better and safer handling of the
data. Also, it fixes the issue that CraftZ sometimes didn't remove a
zombie from the file when it died.
The WorldData class automatically removes all leftover zombies.