Fix auto zombie spawns queueing up
When no player was on the server, CraftZ used to (unintentionally) keep
increasing the number of zombies to spawn. When a player then joined the
game, it was possible that there were dozens or even hundreds of zombies
all over the place. That was an unfortunate bug; it is being fixed with
this commit.