Internal rewrites
With this commit I am trying to push CraftZ into a direction that is
more flexible, extendable, and hopefully easier to understand and
manage.
The whole system is based on Module classes which each represent a
single feature or sub-system. There are no listener classes anymore,
rather each Module listens to the events it needs to process.
Also, I am slowly trying to eliminate static access and to make more use
of instances.

As basically all of the code was moved and sometimes rewritten, bugs are
likely to occur. Please report as usual.