Add backpacks
This commit adds backpacks. They are physical items; chests are used for
display.
You can obtain a backpack like a normal item via kits or loot chests,
and there are by default already some in both.
- To add them to kits, use '/craftz makebackpack <size> <name>' (the
name can contain spaces), this will create an item that you can place in
your inventory.
- To add them to lootchests, use the syntax '<backpack:size:name>', for
example '<backpack:18:British Assault Pack>'.
Note that backpack sizes must be multiples of 9 (Minecraft limitation).
It is impossible to have more than two backpacks at the same time, or to
place a backpack in a backpack. They will keep their items forever, even
when dropped on the floor or given to another player.
It could be possible that they produce a bit of data trash over time,
haven't yet found a way to prevent this.