Remove rocket launcher
This commit removes the 'rocket launcher' module.
That feature has been in CraftZ for a little while, although very few
people even know it exists. Due to that and the issues it has caused
(e.g. TNT not causing any block damage), it is now considered to be of
too little usefulness.